Add tests & run pre-commit
